+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - `--help` or `-h` flag for help information
 ### Changed
 - `-v` flag changed to `-V`
+- User is notified of unrecognised command line arguments and gex exits instead of quietly ignoring
 - Status says "Unstaged changes" and "Staged changes" instead of "files"
 ### Fixed
 - gex not recognising git repositories from within subdirectories ([#2](https://github.com/Piturnah/gex/issues/2))