+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - UI to display available commit commands
 - `--help` or `-h` flag for help information
 ### Changed
-- Use <kbd>c c</kbd> to create a commit
+- Use <kbd>c</kbd> <kbd>c</kbd> to create a commit
 - `-v` flag changed to `-V`
 - User is notified of unrecognised command line arguments and gex exits instead of quietly ignoring
 - Status says "Unstaged changes" and "Staged changes" instead of "files"