+++ CHANGELOG.md
@@ -11,11 +11,12 @@
 - Colouring of `+` and `-` in stdout propagation
 - `--help` or `-h` flag for help information
 ### Changed
-- Propagate all of stdout instead of only first line
 - Use <kbd>c</kbd> <kbd>c</kbd> to create a commit
 - `-v` flag changed to `-V`
 - User is notified of unrecognised command line arguments and gex exits instead of quietly ignoring
 - Status says "Unstaged changes" and "Staged changes" instead of "files"
+- Propagate all of stdout instead of only first line
+
 ### Fixed
 - Showing empty stdout or stderr in the case that the exit code didn't match 
 - gex not recognising git repositories from within subdirectories ([#2](https://github.com/Piturnah/gex/issues/2))